# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar archivos, directorios y etiquetas en Nostalgia.",
        "title": "Nostalgia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/scans/recent": {
            "get": {
                "description": "Devuelve una lista de los escaneos realizados recientemente",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "Obtener escaneos recientes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Cantidad máxima de resultados",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.NScan"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tags": {
            "get": {
                "description": "Devuelve la lista completa de etiquetas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Obtener todas las etiquetas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tags/popular": {
            "get": {
                "description": "Devuelve las etiquetas más utilizadas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Obtener etiquetas populares",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tags/search": {
            "get": {
                "description": "Busca archivos asociados a una etiqueta específica",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Buscar archivos por etiqueta",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la etiqueta",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Resultados por página",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.NFileSearchResultDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tags/{tagName}/directories/{directoryId}": {
            "post": {
                "description": "Asocia una etiqueta a un directorio específico y a todos sus archivos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Añadir etiqueta a directorio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre de la etiqueta",
                        "name": "tagName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID del directorio",
                        "name": "directoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.NDirectory": {
            "type": "object",
            "properties": {
                "date_modified": {
                    "type": "string"
                },
                "directory_count": {
                    "type": "integer"
                },
                "file_count": {
                    "type": "integer"
                },
                "file_nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NFileNode"
                    }
                },
                "full_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_source": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parent_directory": {
                    "$ref": "#/definitions/domain.NDirectory"
                },
                "parent_directory_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NTag"
                    }
                }
            }
        },
        "domain.NFile": {
            "type": "object",
            "properties": {
                "date_modified": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NTag"
                    }
                }
            }
        },
        "domain.NFileNode": {
            "type": "object",
            "properties": {
                "directory": {
                    "$ref": "#/definitions/domain.NDirectory"
                },
                "file": {
                    "$ref": "#/definitions/domain.NFile"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "ndirectory_id": {
                    "type": "integer"
                },
                "nfile_id": {
                    "type": "integer"
                },
                "nscan_id": {
                    "type": "integer"
                },
                "scan": {
                    "$ref": "#/definitions/domain.NScan"
                }
            }
        },
        "domain.NScan": {
            "type": "object",
            "properties": {
                "date_created": {
                    "type": "string"
                },
                "directory_count": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "existing_file_repeated_count": {
                    "type": "integer"
                },
                "file_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "internal_file_repeated_count": {
                    "type": "integer"
                },
                "root_directory": {
                    "$ref": "#/definitions/domain.NDirectory"
                },
                "root_directory_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.NScanStatus"
                }
            }
        },
        "domain.NScanStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "NScanStatusNotStarted",
                "NScanStatusInProgress",
                "NScanStatusCompleted",
                "NScanStatusFailed"
            ]
        },
        "domain.NTag": {
            "type": "object",
            "properties": {
                "directories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NDirectory"
                    }
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NFile"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.NDirectoryDto": {
            "type": "object",
            "required": [
                "full_path",
                "name"
            ],
            "properties": {
                "date_modified": {
                    "type": "string"
                },
                "directories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NDirectoryDto"
                    }
                },
                "directory_count": {
                    "type": "integer"
                },
                "file_count": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NFileDto"
                    }
                },
                "full_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_directory_id": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "size_raw": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.NFileDto": {
            "type": "object",
            "required": [
                "hash",
                "name",
                "path"
            ],
            "properties": {
                "date_modified": {
                    "type": "string"
                },
                "extension": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_directory": {
                    "$ref": "#/definitions/dto.NDirectoryDto"
                },
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "size_raw": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.NFileSearchResultDto": {
            "type": "object",
            "properties": {
                "contains": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NFileDto"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}